pyproject.toml
[tool.poetry]
name = "frinx-workers-boilerplate"
version = "6.1.0"
description = ""
authors = ["Jozef Volak <jvolak@frinx.io>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
frinx-inventory-worker = {git = "https://github.com/FRINXio/frinx-services-python-workers.git", tag = "6.1.0", subdirectory = "inventory/python"}
frinx-schellar-worker = {git = "https://github.com/FRINXio/frinx-services-python-workers.git", tag = "6.1.0", subdirectory = "schellar/python"}
frinx-uniconfig-worker = {git = "https://github.com/FRINXio/frinx-services-python-workers.git", tag = "6.1.0", subdirectory = "uniconfig/python"}
frinx-resource-manager-worker = {git = "https://github.com/FRINXio/frinx-services-python-workers.git", tag = "6.1.0", subdirectory = "resource-manager/python"}
frinx-topology-discovery-worker = {git = "https://github.com/FRINXio/frinx-services-python-workers.git", tag = "6.1.0", subdirectory = "topology-discovery/python"}
frinx-http-worker = {git = "https://github.com/FRINXio/frinx-services-python-workers.git", tag = "6.1.0", subdirectory = "misc/python/http"}
frinx-kafka-worker = {git = "https://github.com/FRINXio/frinx-services-python-workers.git", tag = "6.1.0", subdirectory = "misc/python/kafka"}
frinx-influxdb-worker = {git = "https://github.com/FRINXio/frinx-services-python-workers.git", tag = "6.1.0", subdirectory = "misc/python/influxdb"}
frinx-conductor-system-test-worker = {git = "https://github.com/FRINXio/frinx-services-python-workers.git", tag = "6.1.0", subdirectory = "misc/python/conductor-system-test"}
frinx-python-lambda = {git = "https://github.com/FRINXio/frinx-services-python-workers.git", tag = "6.1.0", subdirectory = "misc/python/python-lambda"}

[tool.poetry.group.dev.dependencies]
ruff = "^0"
mypy = "^1"
pyright = "^1"

[tool.pyright]
reportGeneralTypeIssues = "none"
reportArgumentType = "none"
reportIncompatibleVariableOverride = false

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
extend-select = ['Q', 'RUF100', 'C90', 'UP', 'I', 'E', 'F', 'N', 'PL']
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]

[tool.ruff.lint.isort]
known-first-party = ['loopback_interface', 'tests']
force-single-line = true

[tool.ruff.lint.flake8-quotes]
inline-quotes = 'double'
multiline-quotes = 'double'

[tool.ruff.lint.mccabe]
max-complexity = 14

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 10

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod", "pydantic.field_validator", "pydantic.model_validator"]

[tool.mypy]
python_version = "3.10"
plugins = ["pydantic.mypy"]
explicit_package_bases = true
namespace_packages = true
strict = true
pretty = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_subclassing_any = false
warn_unused_ignores = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
